refactor: move the schema definitions folder
